FrameSent callback to track out frames
Added FrameSent callback in the AMQP client core.  Is called whenever a frame is sent; can be used to track exact moment messages are sent out on the wire.

Refactored the examples/ to use example.pm:init() which can provide callbacks at runtime.

Added examples of FrameSent being used to do a simple 1-N publish and exit.  Call enqueue.pl with a list of messages, and dequeue.pl with the number of messages to wait for from the test queue.